requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.12
aiosignal==1.3.2
anyrun-sdk==1.2.3
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.7.0
exceptiongroup==1.2.2
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
Pygments==2.19.1
pytest==8.3.5
pytest-cov==6.0.0
python-dotenv==1.0.1
requests==2.31.0
rich==13.9.4
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
